+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Add ability to loop the currently playing song
 - Add ability to loop entire session
 
+* Add now playing indicator in song library
+
 ### Changed
 
 - Improve tack name handling