+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add ability to play all the songs in the session continuously
 - Add ability to loop the currently playing song
 - Add ability to loop entire session
+- Add now playing indicator in song library
 
 ### Changed